+++ packages/ws/src/index.ts
@@ -1,2 +1,3 @@
 export * from './configs';
 export * from './cli';
+export { compile as buildTranslations } from './lib/i18n';